Added a missing requirement. Changed the datatype of a tensor to float.

+++ src/embbeded_datasets.py
@@ -30,8 +30,8 @@ class embbededDataset:
 
     def get_train_data(self):
         train_codes = torch.Tensor(torch.load(os.path.join(self.dataset_loc, "train_codes.pt")))
-        labels = torch.load(os.path.join(self.dataset_loc, "train_labels.pt"))
-        
+        labels = torch.load(os.path.join(self.dataset_loc, "train_labels.pt")).float()
+
         train_labels = torch.Tensor(labels).cpu().float()
 
         if self.transform:
